DB: use local time for last_caom_mod
This is because MySQL's automatic timestamp update feature uses local
time.  Since we are now using that for "last_modified", which we
compare to "last_caom_mod", this commit alters "last_caom_mod" also
to be written in local time.  (If we revert to using triggers to
update "last_modified" then we could also revert to using UTC
for these timestamps.  The important thing is that they be in the
same time zone!)