Avoid returning None when tuple expected
The obs_bounds_to_moc function should return a (MOC, set) tuple
when called with project_info=True.  However it previously simply
returned None if no cells were found.  It now always returns a
tuple for project_info=True so that it is safe to unpack the results
into a pair of variables.